Store: switch from Adler32 checksum to CRC32.
Adler has fundamental weakness for small chunks:
https://en.wikipedia.org/wiki/Adler-32#Weakness
Jonathan Stone discovered in 2001 that Adler-32 has a weakness for very short messages. He wrote "Briefly, the problem is that, for very short packets, Adler-32 is guaranteed to give poor coverage of the available bits. Don't take my word for it, ask Mark Adler.  :-)"[4] The problem is that sum A does not wrap for short messages. The maximum value of A for a 128-byte message is 32640, which is below the value 65521 used by the modulo operation. An extended explanation can be found in RFC 3309, which mandates the use of CRC32C instead of Adler-32 for SCTP, the Stream Control Transmission Protocol. Adler-32 has also been shown to be weak for small incremental changes,[5] and also weak "for numbers".[6]