Add methods to specify size limits in bytes
The existing methods (expireStoreSize, sizeLimit) with the size limit in gigabytes
as double sometimes felt a bit clunky when specifying limits less than 1 GB.
This commit adds methods to specify these limits in bytes as a long value.

In order to not break backward-compatibility the methods for specifying the size
limits in bytes have a suffix "Bytes", e. g. `DBMaker.sizeLimitBytes(long)` instead
of overloading the original methods, e. g. `DBMaker.sizeLimit(double)`.