Volume: Set `closed` on constructor exception
If an IOException is thrown during construction (e.g. when
the target directory does not exist or is not writeable),
the `closed` field is never set to true even though the
caller will not receive an instance on which it can call
`close()` later. This periodically leads to the error
message:

```
    Open Volume was GCed, possible file handle leak.
```